backend/requirements.txt
django==5.1.6                     # A high-level Python web framework that encourages rapid development and clean, pragmatic design
    django-cors-headers~=4.7.0    # A Django App that adds Cross-Origin Resource Sharing (CORS) headers to responses
    django-environ~=0.12.0        # Twelve-factor methodology to configure your app
    djangorestframework~=3.15.2   # Web APIs for Django, made easy
